MM-18509: Revamp logging architecture (#66)
* MM-18509: Revamp logging architecture

With this PR, we attempt to completely overhaul the logging mechanism
of the push proxy. There were 2 primary shortcomings which are addressed:
- Config struct was a global variable shared amongst the entire codebase.
- Log methods were global functions which were being called from everywhere.

To fix this,
- We create a server struct and convert the start and stop into methods.
- Pass the config struct into this server struct, so that all state is localized.

For file logging, we use the lumberjack writer which automatically takes care of
file rotation which is a nice thing to have.
By default, if the fields are not set (which would be the case if the config file
wasn't changed), the default behavior of only logging to console would be maintained.

Other notable improvements are:
- Fix a small typo.
- Remove global signal channel.

Overall, this reduces the public API surface by a lot. And adds an improved logging API
with dedicated formatter methods to avoid creating ugly fmt.Sprintf strings.

* Fix linter

* Create a default log directory